+++ libopendmi/src/backend/freebsd.c
@@ -18,6 +18,7 @@
 #include <opendmi/entry.h>
 #include <opendmi/utils.h>
 
+#include <opendmi/backend/generic.h>
 #include <opendmi/backend/freebsd.h>
 
 typedef struct dmi_freebsd_session dmi_freebsd_session_t;
@@ -37,18 +38,6 @@ static bool dmi_freebsd_close(dmi_context_t *context);
 static void dmi_freebsd_session_free(dmi_freebsd_session_t *session);
 
 static bool dmi_freebsd_get_entry_addr(dmi_context_t *context, size_t *paddr);
-
-#if defined(__i386__) or defined(__x86_64__)
-    static bool dmi_freebsd_find_entry_addr(dmi_context_t *context, size_t *paddr);
-
-    static bool dmi_freebsd_find_anchor(
-            dmi_context_t *context,
-            dmi_data_t    *buffer,
-            size_t         base_addr,
-            size_t         area_size,
-            const char    *anchor,
-            size_t        *paddr);
-#endif
 
 dmi_backend_t dmi_freebsd_backend =
 {
@@ -92,7 +81,7 @@ static dmi_data_t *dmi_freebsd_read_entry(dmi_context_t *context, size_t *plengt
         found = dmi_freebsd_get_entry_addr(context, &addr);
 #       if defined(__i386__) || defined(__x86_64__)
             if (not found)
-                found = dmi_freebsd_find_entry_addr(context, &addr);
+                found = dmi_generic_find_entry_addr(context, DMI_FREEBSD_DEV_MEMORY, &addr);
 #       endif
 
         if (not found) {
@@ -161,7 +150,6 @@ static void dmi_freebsd_session_free(dmi_freebsd_session_t *session)
 static bool dmi_freebsd_get_entry_addr(dmi_context_t *context, size_t *paddr)
 {
     char str[KENV_MVALLEN + 1];
-    unsigned long long addr;
 
     assert(context != nullptr);
     assert(paddr   != nullptr);
@@ -177,81 +165,5 @@ static bool dmi_freebsd_get_entry_addr(dmi_context_t *context, size_t *paddr)
 		return false;
 	}
 
-    errno = 0;
-    addr  = strtoull(str, NULL, 0);
-
-    if (((errno == ERANGE) and (addr == ULLONG_MAX)) or (addr > SIZE_MAX)) {
-        dmi_error_raise_ex(context, DMI_ERROR_SYSTEM, "SMBIOS address is out of range: %s", str);
-        return false;
-    }
-
-    dmi_log_debug(context, "Found SMBIOS address: 0x%zx", addr);
-    *paddr = addr;
-
-    return true;
+    return dmi_generic_parse_entry_addr(context, str, paddr);
 }
-
-#if defined(__i386__) or defined(__x86_64__)
-static bool dmi_freebsd_find_entry_addr(dmi_context_t *context, size_t *paddr)
-{
-    const size_t base_addr = 0xF0000;
-    const size_t area_size = 0x10000;
-
-    dmi_data_t *buffer = nullptr;
-    bool        found  = false;
-
-    dmi_log_debug(context, "Running memory scan...");
-
-    buffer = dmi_memory_get(context, DMI_FREEBSD_DEV_MEMORY, base_addr, area_size);
-    if (buffer == nullptr)
-        return false;
-
-    found =
-        dmi_freebsd_find_anchor(context, buffer, base_addr, area_size, DMI_ANCHOR_V30, paddr) or
-        dmi_freebsd_find_anchor(context, buffer, base_addr, area_size, DMI_ANCHOR_V21, paddr) or
-        dmi_freebsd_find_anchor(context, buffer, base_addr, area_size, DMI_ANCHOR_LEGACY, paddr);
-
-    if (not found)
-        dmi_log_debug(context, "No SMBIOS entry point found");
-
-    dmi_free(buffer);
-
-    return found;
-}
-
-static bool dmi_freebsd_find_anchor(
-        dmi_context_t *context,
-        dmi_data_t    *buffer,
-        size_t         base_addr,
-        size_t         area_size,
-        const char    *anchor,
-        size_t        *paddr)
-{
-    size_t length;
-    size_t offset;
-
-    assert(context != nullptr);
-    assert(buffer != nullptr);
-    assert(base_addr % 16 == 0);
-    assert((area_size >= 16) and (area_size % 16 == 0));
-    assert(anchor != nullptr);
-    assert(paddr != nullptr);
-
-    length = strlen(anchor);
-    assert(length <= 16);
-
-    dmi_log_debug(context, "Scanning for SMBIOS anchor: '%s'...", anchor);
-
-    for (offset = 0; offset <= area_size - DMI_ENTRY_MAX_SIZE; offset += 16) {
-        if (memcmp(buffer + offset, anchor, length) == 0) {
-            *paddr = base_addr + offset;
-            dmi_log_debug(context, "Found SMBIOS address: 0x%zx", *paddr);
-            return true;
-        }
-    }
-
-    dmi_log_debug(context, "No SMBIOS anchor found");
-
-    return false;
-}
-#endif