+++ libopendmi/src/backend/linux.c
@@ -82,11 +82,13 @@ static bool dmi_linux_open(dmi_context_t *context, const char *path)
         if (asprintf(&table_path, "%s/%s", dmi_linux_sysfs_path, dmi_linux_table_file) < 0)
             break;
 
-        session->entry = dmi_file_get(context, entry_path, &session->entry_size);    
+        session->entry_size = 0;
+        session->entry = dmi_file_get(context, entry_path, -1, &session->entry_size);    
         if (session->entry == nullptr)
             break;
 
-        session->table = dmi_file_get(context, table_path, &session->table_size);
+        session->table_size = 0;
+        session->table = dmi_file_get(context, table_path, -1, &session->table_size);
         if (session->table == nullptr)
             break;